Implementado funcionalidade relacionado ao objeto Conteudo

+++ Backend/wichstream/src/main/java/com/projetointegrador/wichstream/repository/ConteudoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projetointegrador.wichstream.model.Conteudo;
 
-public interface ConteudoRepository extends JpaRepository<Conteudo, Long> {
+public interface ConteudoRepository extends JpaRepository<Conteudo, Integer> {
 
 }
